+++ packages/util/src/bn/sqrt.ts
@@ -4,15 +4,26 @@
 import BN from 'bn.js';
 
 import { assert } from '../assert';
-import { BN_ONE, BN_TWO, BN_ZERO } from '../bn';
+import { BN_MAX_INTEGER, BN_ONE, BN_TWO, BN_ZERO } from '../bn';
 
-// https://golb.hplar.ch/2018/09/javascript-bigint.html
-function newtonIteration (n: BN, x0: BN): BN {
-  const x1 = n.div(x0).iadd(x0).ishrn(1);
+// https://stackoverflow.com/questions/53683995/javascript-big-integer-square-root/
+function newtonSqrt (n: BN): BN {
+  // shortcut < 2^53 -1 to use the JS utils
+  if (n.lte(BN_MAX_INTEGER)) {
+    return new BN(Math.floor(Math.sqrt(n.toNumber())));
+  }
 
-  return (x0.eq(x1) || x0.eq(x1.sub(BN_ONE)))
-    ? x0
-    : newtonIteration(n, x1);
+  let x0 = new BN(1);
+
+  while (true) {
+    const x1 = n.div(x0).iadd(x0).ishrn(1);
+
+    if (x0.eq(x1) || x0.eq(x1.sub(BN_ONE))) {
+      return x0;
+    }
+
+    x0 = x1;
+  }
 }
 
 /**
@@ -33,5 +44,5 @@ export function bnSqrt (value: BN): BN {
 
   return value.lt(BN_TWO)
     ? value
-    : newtonIteration(value, new BN(1));
+    : newtonSqrt(value);
 }